backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.7
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
isort==5.12.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
pep8-naming==0.13.3
pycodestyle==2.11.0
pyflakes==3.1.0
pyproject-toml==0.0.10
python-dotenv==1.0.0
pytz==2023.3
referencing==0.30.2
rpds-py==0.10.2
sqlparse==0.4.4
toml==0.10.2
pillow==10.0.1
typing_extensions==4.7.1
tzdata==2023.3
PyYAML==6.0.1
drf-spectacular==0.26.4
inflection==0.5.1
uritemplate==4.1.1
djoser==2.2.0
django-rest-framework-social-oauth2==1.1.0
django-oauth-toolkit==2.3.0
djangorestframework>=3.10
django-cors-headers==4.2.0
psycopg2-binary==2.9.9
gunicorn==20.1.0
drf-extra-fields==3.7.0
drf-standardized-errors==0.12.5
pytest-django
